# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Noroff EP e-commerce",
    "description": "APIs and a database to showcase their products for customers to purchase. Admin front-end, which uses the back-end system to manage the data."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "summary": "Fetch all brands",
        "description": "Fetches all brands from the database",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BrandsFetched"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "summary": "Create a new brand",
        "description": "Create a new brand. Requires admin authentication",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Brand name to be created",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateBrand"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/BrandCreated"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/brands/{id}": {
      "put": {
        "tags": [
          "Brands"
        ],
        "summary": "Update a brand",
        "description": "Update a brand name by ID. Requires admin authentication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the brand to update"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated brand name",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateBrand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BrandUpdated"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Brands"
        ],
        "summary": "Delete a brand",
        "description": "Delete a brand by ID. Requires admin authentication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BrandDeleted"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Fetch all categories",
        "description": "Fetches all categories from the database",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CategoriesFetched"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Create a new category",
        "description": "Create a new category. Requires admin authentication",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Category name to be created",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateCategory"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CategoryCreated"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/categories/{id}": {
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Update a category",
        "description": "Update a category name by ID. Requires admin authentication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the category to update"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated category name",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CategoryUpdated"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Delete a category",
        "description": "Delete a category by ID. Requires admin authentication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CategoryDeleted"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/orders/": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Fetch the logged-in user",
        "description": "A list of the logged-in user",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserOrder"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/orders/{id}/": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Update status of an order",
        "description": "Update the status of an order, only options available are \\'In Progress\\', \\'Ordered\\', \\'Completed\\'. Requires Admin authentication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the order to update"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Update status to an order",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UpdateOrder"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "BrandsFetched": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Viewing all brands"
            },
            "brands": {
              "type": "array",
              "example": [
                "List of all brands"
              ],
              "items": {
                "type": "string"
              }
            }
          },
          "required": [
            "result",
            "brands"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "CreateBrand": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Microsoft"
        }
      },
      "required": [
        "name"
      ]
    },
    "BrandCreated": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Created new brand"
            },
            "brand": {
              "type": "string",
              "example": "Microsoft"
            }
          },
          "required": [
            "result",
            "brand"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "UpdateBrand": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Amazon"
        }
      },
      "required": [
        "name"
      ]
    },
    "BrandUpdated": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Brand is updated"
            },
            "brand": {
              "type": "string",
              "example": "Amazon"
            }
          },
          "required": [
            "result",
            "brand"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "BrandDeleted": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "array",
              "example": [
                "Brand is deleted"
              ],
              "items": {
                "type": "string"
              }
            }
          },
          "required": [
            "result"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "CategoriesFetched": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Viewing all categories"
            },
            "categories": {
              "type": "array",
              "example": [
                "List of all categories"
              ],
              "items": {
                "type": "string"
              }
            }
          },
          "required": [
            "result",
            "categories"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "CreateCategory": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Monitors"
        }
      },
      "required": [
        "name"
      ]
    },
    "CategoryCreated": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Created new category"
            },
            "category": {
              "type": "string",
              "example": "Monitors"
            }
          },
          "required": [
            "result",
            "category"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "UpdateCategory": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Keyboards"
        }
      },
      "required": [
        "name"
      ]
    },
    "CategoryUpdated": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Category is updated"
            },
            "category": {
              "type": "string",
              "example": "Keyboards"
            }
          },
          "required": [
            "result",
            "category"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "CategoryDeleted": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "array",
              "example": [
                "Category is deleted"
              ],
              "items": {
                "type": "string"
              }
            }
          },
          "required": [
            "result"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
...
    "UserOrder": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Viewing all orders from user:"
            },
            "orders": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "OrderNumber": {
                    "type": "string",
                    "example": "j8a72g31"
                  },
                  "Status": {
                    "type": "string",
                    "example": "In Progress"
                  },
                  "MembershipStatus": {
                    "type": "string",
                    "example": "Bronze"
                  },
                  "TotalPrize": {
                    "type": "string",
                    "example": "650.00"
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "Purchase date"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "Last status update"
                  },
                  "UserId": {
                    "type": "number",
                    "example": 2
                  },
                  "OrderItems": {
                    "type": "array",
                    "example": [
                      "List of products ordered"
                    ],
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "id",
                  "OrderNumber",
                  "Status",
                  "MembershipStatus",
                  "TotalPrize",
                  "createdAt",
                  "updatedAt",
                  "UserId",
                  "OrderItems"
                ]
              }
            }
          },
          "required": [
            "result",
            "orders"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
...
    "OrderStatus": {
      "type": "object",
      "properties": {
        "status": {
          "type": "array",
          "example": [
            "In Progress",
            "Ordered",
            "Completed"
          ],
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "status"
      ]
    },
    "UpdateOrder": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Order has been updated"
            },
            "orders": {
              "type": "array",
              "example": [
                "In Progress",
                "Ordered",
                "Completed"
              ],
              "items": {
                "type": "string"
              }
            }
          },
          "required": [
            "result",
            "orders"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
...
  }
}